modifié :         makefile 	modifié :         src/basic_functions/power.c 	modifié :         src/main.c

+++ src/basic_functions/power.c
@@ -7,12 +7,12 @@ float pow_like(float *number, float *power) {
         result_pow = 1.0;
     }
     else if  (*power > 0.0) {
-        for (int iterator = 1; (iterator < abs(*power)); iterator++) {
+        for (int iterator = 1; (iterator < abs((int) *power)); iterator++) {
             result_pow *= *number;
         }
     }
     else if (*power < 0.0) {
-        for (int iterator = 1; (iterator < abs(*power) + 2.0); iterator++) {
+        for (int iterator = 1; (iterator < abs((int) *power) + 2.0); iterator++) {
             result_pow /= *number;
         }
     }
